productos/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1  
pydantic==2.8.2
pydantic-settings==2.3.4
sqlalchemy==2.0.16
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psycopg2-binary==2.9.9